improved comments
