+++ jsconfig.json
@@ -1,0 +1,5 @@
+{
+	"compilerOptions": {
+		"target": "ES6"
+	}
+}
